pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "selenium_web_automation_tools"
dynamic = ["version"]
description = "A Selenium-based web automation toolkit using PocketFlow."
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "selenium",
    "webdriver-manager",
    "pyyaml",
    "pocketflow",
]

[tool.setuptools.dynamic]
version = {attr = "selenium_web_automation_tools.__version__"}

[project.urls]
"Homepage" = "https://github.com/yourusername/selenium_web_automation_tools"
"Bug Tracker" = "https://github.com/yourusername/selenium_web_automation_tools/issues"

[tool.setuptools.packages.find]
include = ["selenium_web_automation_tools*"]
